✨ feat : AWS 관련 메서드 분할
